Backend

This is an abstract class that serves as a base class for all concrete backend implementations. It defines the common properties that all concrete backends require.

Details

This class cannot be instantiated. It needs to be extended by concrete subclasses that implement the pure virtual methods. Instances of concrete backend implementations can be conveniently obtained using the BackendFactory class.

See also

Super class

parabar::Service -> Backend

Active bindings

cluster

The cluster object used by the backend. For SyncBackend objects, this is a cluster object created by parallel::makeCluster(). For AsyncBackend objects, this is a permanent R session created by callr::r_session that contains the parallel::makeCluster() cluster object.

supports_progress

A boolean value indicating whether the backend implementation supports progress tracking.

active

A boolean value indicating whether the backend implementation has an active cluster.

Methods

Public methods

Backend$new()

Backend$clone()

Inherited methods

parabar::Service$clear()

parabar::Service$evaluate()

parabar::Service$export()

parabar::Service$get_output()

parabar::Service$peek()

parabar::Service$sapply()

parabar::Service$start()

parabar::Service$stop()

Method new()

Create a new Backend object.

Usage

Backend$new()

Returns

Instantiating this call will throw an error.

Method clone()

The objects of this class are cloneable with this method.

Usage

Backend$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.