Package Exceptions

This class contains static methods for throwing exceptions with informative messages.

Format

Exception$abstract_class_not_instantiable(object)

Exception for instantiating abstract classes or interfaces.

Exception$method_not_implemented()

Exception for calling methods without an implementation.

Exception$feature_not_developed()

Exception for running into things not yet developed.

Exception$not_enough_cores()

Exception for requesting more cores than available on the machine.

Exception$cluster_active()

Exception for attempting to start a cluster while another one is active.

Exception$cluster_not_active()

Exception for attempting to stop a cluster while not active.

Exception$async_task_not_started()

Exception for reading results while an asynchronous task has not yet started.

Exception$async_task_running()

Exception for reading results while an asynchronous task is running.

Exception$async_task_completed()

Exception for reading results while a completed asynchronous task has unread results.

Exception$async_task_error(error)

Exception for errors while running an asynchronous task.

Exception$temporary_file_creation_failed()

Exception for reading results while an asynchronous task is running.

Exception$type_not_assignable(actual, expected)

Exception for when providing incorrect object types.

Exception$unknown_package_option(option)

Exception for when requesting unknown package options.

Exception$primitive_as_task_not_allowed()

Exception for when decorating primitive functions with progress tracking.

Exception$array_margins_not_compatible(actual, allowed)

Exception for using improper margins in the Service$apply operation.