Package Helpers

This class contains static helper methods.

Format

Helper$get_class_name(object)

Helper for getting the class of a given object.

Helper$is_of_class(object, class)

Check if an object is of a certain class.

Helper$get_option(option)

Get package option, or corresponding default value.

Helper$set_option(option, value)

Set package option.

Helper$check_object_type(object, expected_type)

Check the type of a given object.

Helper$check_array_margins(margins, dimensions)

Helper to check array margins for the BackendService$apply operation.