Service

This is an interface that defines the operations available on a Backend implementation. Backend implementations and the Context class must implement this interface.

See also

Methods

Public methods

Service$new()

Service$start()

Service$stop()

Service$clear()

Service$peek()

Service$export()

Service$evaluate()

Service$sapply()

Service$get_output()

Service$clone()

Method new()

Create a new Service object.

Usage

Service$new()

Returns

Instantiating this call will throw an error.

Method start()

Start the backend.

Usage

Service$start(specification)

Arguments

specification

An object of class Specification that contains the backend configuration.

Returns

This method returns void. The resulting backend must be stored in the .cluster private field on the Backend abstract class, and accessible to any concrete backend implementations via the active binding cluster.

Method stop()

Stop the backend.

Usage

Service$stop()

Returns

This method returns void.

Method clear()

Remove all objects from the backend. This function is equivalent to calling rm(list = ls(all.names = TRUE)) on each node in the backend.

Usage

Service$clear()

Details

This method is ran by default when the backend is started.

Returns

This method returns void.

Method peek()

Inspect the backend for variables available in the .GlobalEnv.

Usage

Service$peek()

Returns

This method returns a list of character vectors, where each element corresponds to a node in the backend. The character vectors contain the names of the variables available in the .GlobalEnv on each node.

Method export()

Export variables from a given environment to the backend.

Usage

Service$export(variables, environment)

Arguments

variables

A character vector of variable names to export.

environment

An environment object from which to export the variables.

Returns

This method returns void.

Method evaluate()

Evaluate an arbitrary expression on the backend.

Usage

Service$evaluate(expression)

Arguments

expression

An unquoted expression to evaluate on the backend.

Returns

This method returns the result of the expression evaluation.

Method sapply()

Run a task on the backend akin to parallel::parSapply().

Usage

Service$sapply(x, fun, ...)

Arguments

x

A vector (i.e., usually of integers) to pass to the fun function.

fun

A function to apply to each element of x.

...

Additional arguments to pass to the fun function.

Returns

This method returns void. The output of the task execution must be stored in the private field .output on the Backend abstract class, and is accessible via the get_output() method.

Method get_output()

Get the output of the task execution.

Usage

Service$get_output()

Details

This method fetches the output of the task execution after calling the sapply() method. It returns the output and immediately removes it from the backend. Therefore, subsequent calls to this method are not advised. This method should be called after the execution of a task.

Returns

A vector or list of the same length as x containing the results of the fun. It resembles the format of base::sapply().

Method clone()

The objects of this class are cloneable with this method.

Usage

Service$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.